client: change "is_delay" parameter in check_caps to "no_delay"
Currently, when "is_delay" is true, we send the caps immediately. When
it's false, we delay them. This seems backward to me and is highly
confusing.

Sage says:

"It's definitely confusing.  the call site that motivated the naming is
 the one from tick(), that passes true ... for the delayed flush that
 forces them to be sent but lots of other callers set it to true to avoid
 delaying. It could use a rename/cleanup."

Let's change the name to "no_delay" to indicate that the caller wants
to send caps immediately instead of delaying them.

Signed-off-by: Jeff Layton <jlayton@redhat.com>