client/mds: add flags field to MClientCaps
...and encode/decode it appropriately.

The idea of this field is to be advisory, to allow the MDS to handle
things differently if it so chooses. We deliberately do _not_ offer
any firm policy here.

We start with a flag that tells the MDS when an application may end up
blocking on the result of this cap request.

A new "sync" arg is added to send_cap, and we set the new flag in the
cap request based on its value. For now, the callers all set the sync
boolean to false everywhere to preserve the existing behavior.

Signed-off-by: Jeff Layton <jlayton@redhat.com>